requirements.txt
python-telegram-bot>=21.0
python-dotenv>=1.0.0
pymongo>=4.7.0
dnspython>=2.6.0
requests>=2.32.0
